+++ internal/domain/model/message.go
@@ -8,12 +8,12 @@ import (
 
 type Message struct {
 	gorm.Model
-	ID         uint64    `gorm:"primaryKey"`
-	ChatId     uint64    `json:"chatid,omitempty"`
+	ID         uint      `gorm:"primaryKey"`
+	ChatId     uint      `json:"chatid,omitempty"`
 	Chat       Chat      `gorm:"foreignKey:ChatId"`
-	SenderId   uint64    `json:"senderid,omitempty"`
+	SenderId   uint      `json:"senderid,omitempty"`
 	Sender     User      `gorm:"foreignKey:SenderId"`
-	ReceiverId uint64    `json:"receiverid,omitempty"`
+	ReceiverId uint      `json:"receiverid,omitempty"`
 	Receiver   User      `gorm:"foreignKey:ReceiverId"`
 	Content    string    `json:"content,omitempty"`
 	CreatedAt  time.Time `json:"created_at,omitempty"`
